24_4_4
봄버맨

+++ GrapperStart/Assets/1_Scripts/Player/ScriptableObject/PlayerData.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Player Data")]
 public class PlayerData : ScriptableObject
 {
-    public int playerHp = 10;
+    public float playerHp = 10;
     public float curSpeed = 8;
 
-    public void DamagedHp(int amount)
+    public void DamagedHp(float amount)
     {
         playerHp -= amount;
     }
